backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
google-generativeai==0.8.0
pydantic==2.5.3
kokoro-onnx>=0.4.0
numpy>=1.24.0
scipy>=1.10.0
sounddevice>=0.4.6
